feat(AMS): add conditional form field for startrentaldate and refactor  merge into single component
